[新增]收款模式：常亮收款 [优化]收款模式：副屏显示刷卡核销按钮 [打包]1.6.1
(cherry picked from commit 4101b31aafb3f163c558f75a9e8b24b23d0b8a1a)

+++ app/src/main/java/com/yannuo/dgcanteen/activitys/fragment/FacFragment.kt
@@ -63,7 +63,7 @@ class FacFragment : BaseFragment<FragmentFacBinding>() {
 
             if (payForUI.result == "Y") {
                 bean.apply {
-                    payment = payForUI.actualPayment.ifEmpty { payForUI.payment }.toFloat()
+                    payment = payForUI.actualPayment.ifEmpty { payForUI.payment }
                     custName = payForUI.username
                     accNo = payForUI.accNo
                     acc_bal = payForUI.accBal
